[source-map] Import a benchmark for the mozilla/source-map package.
As requested earlier in the discussion on

  https://github.com/nodejs/benchmarking/issues/138#issuecomment-325573471

this adds a benchmark for the popular source-map package, which covers
both the serialization and the parsing of source maps. The payload is
the source map for the minified source-map.js itself.

We use `backbone-min.map`, `jquery.min.map`, `preact.min.js.map` and
`source-map.min.js.map` (all from the lastest versions) as payloads
for the benchmark.