Add an initial browser version.
This initial page is heavily inspired by the JetStream look, but
otherwise doesn't share anything with it. While JetStream has it's
own custom benchmarking framework, we just rely on benchmark.js
for the browser bundle as well.

Tested in the latest Chrome, Safari and Firefox browsers on
macOS so far.